backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
PyJWT
werkzeug
requests